fix: addressed suggested fixes
