--- core/database/src/desktopMain/kotlin/com/mifos/room/utils/TypeConverter.kt
@@ -1,14 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/android-client/blob/master/LICENSE.md
- */
-package com.mifos.room.utils
-
-import androidx.room.TypeConverter
-
-actual typealias TypeConverter = TypeConverter
